Refactor Notepad file structure
